Finalize de_whitelist_additions.txt
Removing the relatively infrequent lemmas "kibben", "FLOP", "Schoss", "eis", "ces", "ges", "gis" and "pu".
The high frequency classes of the other lemmas (e.g., "kippen", "Flop", "Schoß", "Eis") that made them potential additions do not actually relate to the meanings of the lemmas, as defined in the respective DWDS articles (where the lemma that had been considered as an addition is the main form).